### 2020-05-29 #2274 @whedon
[PRE REVIEW]: ALVA: An adaptive MATLAB package for layered viscoelastic elastic analysis
**Submitting author:** @asmusskar (<a href="http://orcid.org/0000-0003-3176-791X">Asmus Skar</a>)
**Repository:** <a href="https://github.com/asmusskar/ALVA" target ="_blank">https://github.com/asmusskar/ALVA</a>
**Version:** v1.0.0
