### 2020-06-12 1a1117b
Bug in strain calculation fixed
