### 2020-06-12 60e25e0
New example - strain calculation validation
