### 2020-06-12 551b6be
New validation results for strain included
