+++ validation/validation_let.m
@@ -113,6 +113,28 @@ elseif strcmp(alva.validation,'halfspace_shear') > 0
         750	-0.0026	-0.0025	0.0627	0.0016	0.0008	0.0000	0.0012	0.0024	0.2513
         1000	-0.0016	-0.0016	0.0361	0.0007	0.0004	0.0000	0.0007	0.0014	0.1900
         ];
+ elseif strcmp(alva.validation,'halfspace_shear_strain') > 0
+    % format: [eps_x	eps_y	eps_z	eps_xy	eps_yz	eps_xz
+    response = [
+        0	736.6859	389.9417	-210.2096	-9.4578	-0.0006	0.0007
+        10	593.8950	343.2896	-132.3018	1.0855	-4.4535	8.1273
+        20	468.4325	302.7134	-70.5882	4.5071	-9.6577	17.3330
+        30	359.5222	261.2682	-18.5650	4.4650	-11.1794	23.4490
+        40	263.1367	217.2765	27.7279	3.6572	-10.8141	27.1034
+        50	175.8194	171.3471	69.9623	2.8524	-9.7224	29.1766
+        75	-19.1248	52.1867	165.1114	1.5396	-6.4335	30.0798
+        100	-204.0480	-68.9493	257.8612	1.0040	-3.7112	26.4775
+        150	-645.5158	-306.0617	499.9204	1.3688	2.2696	4.2175
+        200	-577.7007	-340.9639	1473.3037	0.9234	42.4749	51.3710
+        250	-514.2167	-347.6736	1287.1622	0.7365	43.0236	40.8279
+        300	-459.5256	-340.5053	1120.9192	0.6007	40.4834	32.9230
+        350	-414.8097	-328.1782	976.9886	0.4867	36.7254	26.9204
+        400	-380.0564	-315.8148	855.1341	0.3921	32.6738	22.2754
+        450	-355.1578	-306.4883	754.4073	0.3168	28.7169	18.6013
+        500	-340.0644	-302.2585	673.8017	0.2592	24.9610	15.5897
+        750	-452.0891	-423.3760	601.8378	0.2236	4.6629	2.6201
+        1000	-300.8785	-289.0444	724.5183	0.0965	13.3008	7.1316
+        ]; 
 else
     error('Validation case not in library')
 end
