### 2020-06-12 b8591cb
New example included
