### 2020-06-12 a4e1bee
Update ALVA_let_validation5.m
