### 2020-06-16 4cc63df
Update ALVA_let_validation1.m
