### 2020-06-16 660b2bb
Update ALVA_let_validation4.m
