+++ README.md
@@ -1,4 +1,4 @@
-Skar, A. and Andersen, S (2020). ALVA. https://doi.org/10.11583/DTU.12387305.v1
+Skar, Asmus and Andersen, Sebastian (2020): Adaptive Layered Viscoelastic Analysis (ALVA). Technical University of Denmark. Software. https://doi.org/10.11583/DTU.12387305.v4
 
 # ALVA
 Adaptive Layered Viscoelastic Analysis (ALVA) is a MATLAB-package for pavement modeling. The aim with the software is to equip the civil engineering community with an advanced pavement modeling tool and computer package that is highly adaptive, transparent and open-access, capable of supporting current and future pavement evaluation needs. 
