### 2020-07-08 #2274 @asmusskar
Hi!

I can recommend Emin Kutay (https://www.egr.msu.edu/people/profile/kutay)
and Sigurdur Erlingsson (
https://www.vti.se/sv/medarbetare/profil/sigurdur_erlingsson/), which to my
knowledge are both top international domain knowledge and Matlab experts. I
can also recommend Morten Herfelt Andersen (mah@optumce.com) a software
developer at OPTUM computational engineering with Matlab skills that kills
;-)

Asmus

On Wed, Jul 8, 2020 at 11:08 AM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @asmusskar <https://github.com/asmusskar> I'm currently looking for
> reviewers for your submission. If you can recommend any please let me know.
>
