### 2020-07-12 #2274 @Kevin-Mattheus-Moerman
@fyangneil @hmhtb would you be interested in reviewing this submission for [JOSS (The Journal of Open Source Software)](https://joss.theoj.org):
_ALVA: An adaptive MATLAB package for layered viscoelastic elastic analysis_

The review takes place on GitHub and focuses on this short paper:
https://github.com/openjournals/joss-papers/blob/joss.02274/joss.02274/10.21105.joss.02274.pdf
and the software:
https://github.com/asmusskar/ALVA

You can tell me if you are interested by replying here. Once I have sufficient reviewers I'll open a dedicated review issue.

If you are unavailable but can recommend alternative reviewers that would be great. Thank you!
