### 2020-07-25 #2274 @Kevin-Mattheus-Moerman
@CBenghi @mdelcambre @satish-annigeri @abhiramm7 @anawaldila would you be interested in reviewing this civil engineering related submission for JOSS (The Journal of Open Source Software):
ALVA: An adaptive MATLAB package for layered viscoelastic elastic analysis

The review takes place on GitHub and focuses on this short paper:
https://github.com/openjournals/joss-papers/blob/joss.02274/joss.02274/10.21105.joss.02274.pdf
and the software:
https://github.com/asmusskar/ALVA

You can tell me if you are interested by replying here. Once I have sufficient reviewers I'll open a dedicated review issue.

If you are unavailable but can recommend alternative reviewers that would be great. Thank you!
