### 2020-07-25 #2274 @AnaWaldila
Yes, I would like to review this paper,
Kind regard,

----------------------------------------------------------------------------------------------------

Ana Waldila de Queiroz Ramiro Reis

Doutoranda em Engenharia Civil - Estrutura (UERJ)

Engenheira Civil

Tel: (21) 9 7945 - 6510

E-mail: anawaldila@hotmail.com

________________________________
De: Kevin Mattheus Moerman <notifications@github.com>
Enviado: sábado, 25 de julho de 2020 15:30
Para: openjournals/joss-reviews <joss-reviews@noreply.github.com>
Cc: Ana Waldila <anawaldila@hotmail.com>; Mention <mention@noreply.github.com>
Assunto: Re: [openjournals/joss-reviews] [PRE REVIEW]: ALVA: An adaptive MATLAB package for layered viscoelastic elastic analysis (#2274)

@CBenghi<https://github.com/CBenghi> @mdelcambre<https://github.com/mdelcambre> @satish-annigeri<https://github.com/satish-annigeri> @abhiramm7<https://github.com/abhiramm7> @AnaWaldila<https://github.com/AnaWaldila> would you be interested in reviewing this civil engineering related submission for JOSS (The Journal of Open Source Software):
ALVA: An adaptive MATLAB package for layered viscoelastic elastic analysis

The review takes place on GitHub and focuses on this short paper:
https://github.com/openjournals/joss-papers/blob/joss.02274/joss.02274/10.21105.joss.02274.pdf
and the software:
https://github.com/asmusskar/ALVA

You can tell me if you are interested by replying here. Once I have sufficient reviewers I'll open a dedicated review issue.

If you are unavailable but can recommend alternative reviewers that would be great. Thank you!
