### 2020-07-27 #2274 @Kevin-Mattheus-Moerman
@CBenghi thanks that would be great. I'll add you for now so we can get started and I'll remind you mid-August. If you have difficulties installing/setting up MATLAB let me know.

@asmusskar can you comment on how compatible this work is with Octave? Does it run in Octave or not really? Are special MATLAB toolboxes required?
