### 2020-07-27 #2274 @asmusskar
Hi.

The package should be fully compatible with Octave. At least this is what I have heard from other users.

Asmus

Sendt fra min iPhone

> Den 27. jul. 2020 kl. 08.46 skrev Kevin Mattheus Moerman <notifications@github.com>:
>
> ﻿
> @CBenghi thanks that would be great. I'll add you for now so we can get started and I'll remind you mid-August. If you have difficulties installing/setting up MATLAB let me know.
>
> @asmusskar can you comment on how compatible this work is with Octave? Does it run in Octave or not really? Are special MATLAB toolboxes required?
>
