### 2020-08-05 #2548 @whedon
[REVIEW]: ALVA: An adaptive MATLAB package for layered viscoelastic analysis
**Submitting author:** @asmusskar (<a href="http://orcid.org/0000-0003-3176-791X">Asmus Skar</a>)
**Repository:** <a href="https://github.com/asmusskar/ALVA" target ="_blank">https://github.com/asmusskar/ALVA</a>
**Version:** v1.2
