### 2020-08-20 #2548 @AnaWaldila
Hi,
I need more 2 weeks, please. My father had an ischemic stroke and I am in hospital with him.
Kind regards,
Ana Reis

Obter o Outlook para Android<https://aka.ms/ghei36>

________________________________
De: Kevin Mattheus Moerman <notifications@github.com>
Enviado: quinta-feira, 20 de agosto de 2020 09:55
Para: openjournals/joss-reviews
Cc: Ana Waldila; Mention
Assunto: Re: [openjournals/joss-reviews] [REVIEW]: ALVA: An adaptive MATLAB package for layered viscoelastic elastic analysis (#2548)

@AnaWaldila<https://github.com/AnaWaldila>, @CBenghi<https://github.com/CBenghi> this is where the review takes place. Please can you let me know when you expect to be able to complete your review? Thanks again for your help! Let me know if you have any questions.
