### 2020-09-02 #2548 @AnaWaldila
Hi Kelvin,
So.... my father deaed in 26th August.
Well, sorry about the long time to reviewer the paper.
I am finishing now. The last part is completing the list of revision that I received from @whedon.
I have a question:

  *   Functionality: Have the functional claims of the software been confirmed?
  *   Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

In Performance, did I put my considerations about the software? Because I have some suggestions for the author of the software. I dind't unsderstand this case.
Well, thank you for understanding and I am sorry for the long time to revision.
Kind Regards,
Ana Reis

----------------------------------------------------------------------------------------------------

Ana Waldila de Queiroz Ramiro Reis

Doutoranda em Engenharia Civil - Estrutura (UERJ)

Engenheira Civil

Tel: (21) 9 7945 - 6510

E-mail: anawaldila@hotmail.com

________________________________
De: Claudio Benghi <notifications@github.com>
Enviado: terça-feira, 1 de setembro de 2020 06:34
Para: openjournals/joss-reviews <joss-reviews@noreply.github.com>
Cc: Ana Waldila <anawaldila@hotmail.com>; Mention <mention@noreply.github.com>
Assunto: Re: [openjournals/joss-reviews] [REVIEW]: ALVA: An adaptive MATLAB package for layered viscoelastic elastic analysis (#2548)

Sorry to hear that, @AnaWaldila<https://github.com/AnaWaldila>.
I need a little more time to adjust to recent workplace changes, will be able to pick up next week.
Sorry for the delay.
Claudio
