### 2020-09-02 #2548 @Kevin-Mattheus-Moerman
@AnaWaldila our condolences on your loss. Thank you for your continued help here. Again we can be flexible with time so please take as much time as you need, all your help is greatly appreciated.

> In Performance, did I put my considerations about the software? Because I have some suggestions for the author of the software. I dind't unsderstand this case.

Yes please comment here if you have minor comments/issues or open dedicated issues on the software repository for longer comments/issues. Let me know if you are still unsure what to do or if you have more questions.
