### 2020-09-04 #2548 @AnaWaldila
Hi Kevin,
I finished the revision. Thank you soo much for the compreension and thank you for the condolences. The last days was hard for me.
How can I send to you my considerations about the paper?
One more time, thank you for compreension and I are allways avaiable for a new revision to you.
Kind Regards,
Ana Reis

----------------------------------------------------------------------------------------------------

Ana Waldila de Queiroz Ramiro Reis

Doutoranda em Engenharia Civil - Estrutura (UERJ)

Engenheira Civil

Tel: (21) 9 7945 - 6510

E-mail: anawaldila@hotmail.com

________________________________
De: Kevin Mattheus Moerman <notifications@github.com>
Enviado: quarta-feira, 2 de setembro de 2020 07:56
Para: openjournals/joss-reviews <joss-reviews@noreply.github.com>
Cc: Ana Waldila <anawaldila@hotmail.com>; Mention <mention@noreply.github.com>
Assunto: Re: [openjournals/joss-reviews] [REVIEW]: ALVA: An adaptive MATLAB package for layered viscoelastic elastic analysis (#2548)

@AnaWaldila<https://github.com/AnaWaldila> our condolences on your loss. Thank you for your continued help here. Again we can be flexible with time so please take as much time as you need, all your help is greatly appreciated.

In Performance, did I put my considerations about the software? Because I have some suggestions for the author of the software. I dind't unsderstand this case.

Yes please comment here if you have minor comments/issues or open dedicated issues on the software repository for longer comments/issues. Let me know if you are still unsure what to do or if you have more questions.
