### 2020-09-10 #2548 @AnaWaldila
Hi, Kelvin
I opened this e-mail in GitHub and I see all boxes. Thank you for help me. I cheked some boxes and next, here, I will put my considerations, because the author needs to rectify some cases in paper.
All parts in bold are my answers.

Functionality
•	Functionality: Have the functional claims of the software been confirmed? **All of results presents in repository confirmed the functional of the software. One result input by the reviewer presents problems in result (layer frictionless).**
•	Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
- **The implementation of ALVA presents 3 types of interface: Bonded, Slip and Frictionless. When the user change the type to Frictionless (line 65 of the first validation example), the results between ALVA and Analytical Case are different. Please, justify.**
- **Presentating a graphic with 3 types of interface (Bonded, Slip and Frictionless), comparing the solution of ALVA and analytical process for the each case.**
- **Its interesting the author compare in all graphics the results between ALVA and ELLEA1 about all interfaces between layers.**
- **Presentating in all scripts references that used.**
- **In repository, its interesting the author explain the input data used in each example (https://github.com/asmusskar/ALVA)**
- **In repository (https://github.com/asmusskar/ALVA) that’s describe ALVA, the author can put an explanation in paragraph of main.m (where describes the archive main.m) about the validation examples, i. e, the validation examples are the main.m archive.**
- **Suggestion to future work: Create an interface using AppDesigner or an interaction with user in Command Window. Validation examples are good to understand the process of input data, but the user need this type of script to input own data to create an analysis.**

Software paper
•	Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided? **About general explanations, YES.**
•	State of the field: Do the authors describe how this software compares to other commonly-used packages? **NO. It needs a clear explanation about this. The author needs to explain how compareted ALVA with ELLA1, ELLVA and analytical analysis. Also that, explain more about ELLEA1, ELLVA and the analytical methods.
All procession of the user can change parameters (like layers, geometry structure, materials and for exemple) are need presented in context, i.e, a short description to user the software.**
•	Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?  **It’s interesting to use an impersonal expressions (expressions like “we” are not common in publishing papers).**
•	References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?

**NO. Next, references presented in software and not presented in list of references**

**Burmister, D. M. (1945). The general theory of stresses and displacements in layered systems. i. Journal of applied physics, 16(2), 89–94.**
**Huang, Y.H., 2003. Pavement Analysis Design, 2nd Edition, Prentice-Hall, New Jersey.**
**Ioannides, A. M., & Khazanovich, L. (1998). General formulation for multilayered pavement systems. Journal of transportation engineering, 124(1), 82–90.**
**Levenberg, E. (2018). ELLVA_VD: Isotropic Layered Viscoelasticity in Excel: Analysis tool for interpretation of deflections measured with a moving load**

**Next, references presented in paper and not presented in software**

**Chabot, A., Chupin, O., Deloffre, L., & Duhamel, D. (2010). ViscoRoute 2.0 a: Tool for the simulation of moving load effects on asphalt pavement. Road Materials and Pavement Design, 11(2), 227–250. doi:10.1080/14680629.2010.9690274**
**Erlingsson, S., & Ahmed, A. W. (2013). Fast layered elastic response program for the analysis of flexible pavement structures. Road Materials and Pavement Design, 14(1), 196–210. doi:10.1080/14680629.2012.757558**
**Khazanovich, L., & Wang, Q. (2007). MnLayer: High-performance layered elastic analysis program. Transportation Research Record, 2037(2037), 63–75. doi:10.3141/2037-06**
**Levenberg, E., & Skar, A. (2020). Analytic pavement modeling with a fragmented layer. International Journal of Pavement Engineering. Taylor; Francis Ltd.**
**Maina, J. W., De Beer, M., & Matsui, K. (2007). Effects of layer interface slip on the response and performance of elastic multi-layered flexible airport pavement systems. Maintenance and Rehabilitation of Pavements and Technological Control - Proceedings of the 5th International Conference, Mairepav 2007, 145–150.**
**Skar, A., Levenberg, E., Andersen, S., & Andersen, M. B. (2020). Analysis of a moving measurement platform based on line profile sensors for project-level pavement evaluation. Road Materials and Pavement Design, 1–17. doi:10.1080/14680629.2020.1741429**

**Next, the list of references that’s appears in both cases (software and list of references), but the citation are differens:**

**Levenberg, E. (2016a). ELLEA1: Isotropic layered elasticity in excel: Pavement analysis tool for students and engineers.**
**Levenberg, E. (2016b). ELLVA1: Isotropic layered viscoelasticity in excel (moving load): Advanced pavement analysis tool for students and engineers.**
 **Levenberg, E. (2016c). Viscoelastic pavement modeling with a spreadsheet. Proceedings of the Eighth International Conference on Maintenance and Rehabilitation of Pavements (mairepav8), 746–755. doi:10.3850/978-981-11-0449-7-132-cd**

**Please, for all cases, put the same type of references in software and in paper. Also that, it's important to references all scripts (some scripts does not have references).**
