### 2020-09-14 #2548 @CBenghi
@Kevin-Mattheus-Moerman,

Unfortuntely all University staff IT is completely absorbed in the solution of a major incident that has crippled all fundamental services at a critical time of the new academic year, which would be hard in normal times, but unbearable in Covid times.

At the moment I cannot tell when I'll have access to online services to install Matlab; sadly, this has been going on for weeks.

@asmusskar, I would be able to test the software using Octave instead of Matlab, do you think this might be worth an attempt?

Apologies for the delay, this incident is truly unprecented in terms of scale and scope for our university.

Thanks,
Claudio
