### 2020-10-04 #2548 @asmusskar
Hi Kevin,

sure, I can respond within a few days. However I'm not sure in what
format.. I respond to the reviewers comments here, and then implement the
changes in the code/paper/readme accordingly?

Asmus

On Sun, Oct 4, 2020 at 12:51 PM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @asmusskar <https://github.com/asmusskar> we do not have strict
> deadlines, but to avoid loosing track of the reviewers I do recommend
> responding in a timely manor.
>
