### 2020-10-09 54c99e7
Update paper.bib
Updated DOI Levenberg (2020) and corrected reference from Maina (2007) to Maina (2004) wrt interface bonding and software GAMES
