+++ paper.bib
@@ -89,24 +89,31 @@
   author = {Levenberg, Eyal}
 }
 
-@unpublished{Levenberg:2020a,
-  title = {Analytic pavement modeling with a fragmented layer},
+@article{Levenberg:2020a,
+  title = {Analytic pavement modelling with a fragmented layer},
   language = {eng},
   publisher = {Taylor and Francis Ltd.},
   journal = {International Journal of Pavement Engineering},
-  pages = {1-22},
+  pages = {1-13},
   year = {2020},
+  issn = {1477268x, 10298436},
+  doi = {10.1080/10298436.2020.1790559},
   author = {Levenberg, Eyal and Skar, Asmus}
 }
 
-@article{Maina:2007a,
-  title = {Effects of layer interface slip on the response and performance of elastic multi-layered flexible airport Pavement systems},
+@article{maina:2004a,
+  title = {Developing software for elastic analysis of pavement structure responses to vertical and horizontal surface loadings},
   language = {eng},
-  publisher = {Politecnico di Torino (DIMEAS)},
-  journal = {Maintenance and Rehabilitation of Pavements and Technological Control - Proceedings of the 5th International Conference, Mairepav 2007},
-  pages = {145-150},
-  year = {2007},
-  author = {Maina, James W. and De Beer, Morris and Matsui, Kunihito}
+  publisher={SAGE Publications Sage CA: Los Angeles, CA},
+  journal = {Transportation Research Record},
+  volume = {1896},
+  number = {1896},
+  pages = {107-118},
+  year = {2004},
+  issn = {21694052, 03611981},
+  isbn = {0309094895},
+    doi = {10.3141/1896-11},
+  author = {Maina, James W. and Matsui, Kunihito}
 }
 
 @article{Skar:2020a,
