### 2020-10-09 8f4c81b
Update paper.md
- line 23: Correct reference and include software GAMES (maina 2004 in change of maina 2007)
- line 29: new figure showing the pavement system
- line 31: description of LET model parameters
- line 33: description of VE model parameters
- line 35: new figure showing the moving load scheme
- line 37: description of load scheme
