+++ README.md
@@ -12,13 +12,13 @@ The components of the ALVA software (shown in the Figure) is briefly described b
 
 The core algorithm behind this package is based on LET, i.e., the classic formulation for an N-layered half-space, shown in Figure \ref{fig1}.
 
-![Figure 1: N-Layered half-space model\label{fig1}](N_layer.png)
+![Figure 1: N-Layered half-space model\label{fig1}](<img src="images/N_layer.png" width="90%">)
 
 In this model all layers are assumed linear elastic, isotropic, homogeneous, fully bonded, and weightless. The model inputs include Young’s modulus $E_n$, Poisson’s ratio $\nu_n$, and layer thickness $t_n$ (where $n$ denotes the layer number). This model is engaged to calculate the response at any point, A_j$, of interest and for a given set of uniformly distributed circular loadings with load radius, $a$, and pressure $q$). An overview of LET model assumptions and solution procedure is given in [@khazanovich:2007a].
 
 The viscoelastic response is approximated based on the LET calculations utilizing the methodology and load scheme suggested by [@Levenberg:2016c] (see Figure \ref{fig2}). Viscoelastic layers are associated with a creep compliance and model parameters $D_0$ and $D_\infty$, the short and long time compliances (respectively), and shape parameters $\tau_D$ and $n_D$, controlling the transition between $D_0$ and $D_\infty$. 
 
-![Figure 2: Load scheme to simulate movement\label{fig2}](VE_mesh.png)
+![Figure 2: Load scheme to simulate movement\label{fig2}](<img src="images/VE_mesh.png" width="90%">)
 
 The load moves in a straight line from $x=-x_0$ (Start) to $x=x_0$ (End). The travel path is decomposed into $N$ intervals $(i=1,…,N)$, each $\Delta x$ long. The point of response evaluation is indicated in the Figure; this point is located near the middle of the travel path (i.e., $x$-coordinate of zero), at $y$-coordinate $y_0$ and depth $z_0$ below the surface. 
 
