### 2020-10-09 c13f3b6
Update paper.md
