+++ examples/ALVA_bonding_validation1.m
@@ -1,0 +1,251 @@
+% DESCRIPTION:
+% This script tests the implementation of the ALVA LET model, calculating 
+% responses at key locations for a multilayered pavement subjected to a 
+% single load and a dual load respectively utilizing the method proposed. 
+% The results are compared to benchmark results reported by the European 
+% Commission (https://trimis.ec.europa.eu/project/advanced-models-...
+% analytical-design-european-pavement-structures). 
+
+clear all, close all, clc
+
+% -------------------------------------------------------------------------
+% Select response analysis type
+% -------------------------------------------------------------------------
+alva.analysis = 'Full';
+% 1) 'Full'     : Conventional full integration with one-step Richardson
+%                 extrapolation (for improved convergence near surface) 
+%                 applied for evaluation of displacements and stresses
+% 2) 'PolFit'   : Use polynomial fit technique to reduce integral length
+%                 according to Andersen et al. (2018) for evaluation of
+%                 surface displacements
+
+% -------------------------------------------------------------------------
+% Numerical parameters
+% -------------------------------------------------------------------------
+alva.N  = 300;  % Number of Bessel zero points in numerical integration
+alva.n  = 30;   % Number of Gauss points points between zero points.
+
+% -------------------------------------------------------------------------
+% Pavement material properties (minimum two layers required)
+% -------------------------------------------------------------------------
+alva.zi = [260 760];         % Depth of first n-1 layers from the 
+                             % surface [mm]: last z = inf, and should not 
+                             % be added NB: zi(i) > zi(i-1) > z(i-2)...
+alva.E  = [5000 200 50];     % Layer Young's moduli [MPa]
+alva.nu = [0.35 0.40 0.45];  % Layer Poisson's ratio [-]
+
+% -------------------------------------------------------------------------
+% Load configuration - Single case
+% -------------------------------------------------------------------------
+
+alva.q  = [0.7];         % Load pressure [MPa] (uniform vertical pressure)
+alva.a  = [150.8];       % Load radii [mm] (circular load)
+alva.Xl = [0.0 0.0];     % Load positions [mm]: [x1 y1; x2 y2;..xi yi];
+
+% -------------------------------------------------------------------------
+% Location of evaluation points: [x1 y1 z1; x2 y2 z2;..] - Single case
+% -------------------------------------------------------------------------
+alva.Xd = [    0 0 0;     0 0 259.9;     0 0 260.01;     0 0 760.1;   
+           150.8 0 0; 151 0 259.9; 150.8 0 260.01; 150.8 0 760.1
+           ];
+       
+% -------------------------------------------------------------------------
+% Initialize system and get response - Single case
[…]
