### 2020-10-09 b2e2199
Update paper.bib
added reference to benchmark validation results
