+++ paper.bib
@@ -1,3 +1,13 @@
+@report{amadeus:2000a,
+  title = {Advanced models for analytical design of European pavement
+structures},
+  author = {{Belgian Road Research Centre}},
+  organization = {European Commission},
+  address = {Brussels, Belgium},
+  year = {2000},
+  url = {https://trimis.ec.europa.eu/project/advanced-models-analytical-design-european-pavement-structures},
+}
+
 @article{Andersen:2020a,
   title = {Efficient reevaluation of surface displacements in a layered elastic half-space},
   language = {eng},
