+++ paper.bib
@@ -1,6 +1,7 @@
 @report{amadeus:2000a,
   title = {Advanced models for analytical design of European pavement
 structures},
+  note = {Report no. RO-97-SC.2137},
   author = {{Belgian Road Research Centre}},
   organization = {European Commission},
   address = {Brussels, Belgium},
