### 2020-10-09 1573dc2
Update paper.md
Updated with text wrt validation (line 49)
