+++ paper.md
@@ -46,7 +46,7 @@ The viscoelastic response is approximated based on the LET calculations utilizin
 
 The load moves in a straight line from <i>x=-x<sub>0</sub></i> (Start) to <i>x=x<sub>0</sub></i> (End). The travel path is decomposed into <i>N</i> intervals (<i>i=1,…,N</i>), each <i>&Delta;x</i> long. The point of response evaluation <i>A<sub>j</sub></i> is indicated in the Figure; this point is located near the middle of the travel path (i.e., <i>x</i>-coordinate of zero), at <i>y</i>-coordinate <i>y<sub>0</sub></i> and depth <i>z<sub>0</sub></i> below the surface. 
 
-The final computer package Adaptive Layered Viscoelastic Analysis (ALVA) offers a near real-time solution for the history of stress, strain, and displacement inside the system at any point of interest resulting from a moving load. The code have been validated by comparing ALVA responses at several positions within pavement system to analytical solutions and other commonly used software packages, i.e.: <i>(i)</i> for a one-layer model with the analytical solution for an elastic half-space,  <i>(ii)</i> for elastic multilayered systems with the software package ELLEA [@Levenberg:2016a], and <i>(ii)</i> for viscoelastic multilayered systems with the software package ELLVA [@Levenberg:2016b]. Moreover, the code have been validated against benchmark results, covering a range of commonly used software packages, published by the European Commission [@amadeus:2000a]. The generic computational scheme proposed enables users to control any feature of the program, including geometrical and material properties, layer interface bonding, loading conditions and numerical parameters. Thus, ALVA supports advanced analysis, e.g., detailed analysis of surface tire interaction and unconventional axle loads, and enables users to balance numerical efficiency and accuracy. ALVA can be used as is [@Andersen:2020a;@Skar:2020a] or serve as computational kernel in new software; supporting future advances in pavement analysis software and development of model-guided data interpretation schemes.
+The final computer package Adaptive Layered Viscoelastic Analysis (ALVA) offers a near real-time solution for the history of stress, strain, and displacement inside the system at any point of interest resulting from a moving load. The code have been validated by comparing ALVA responses at several positions within pavement system to analytical solutions and other commonly used software packages, i.e.: <i>(i)</i> for a one-layer model with the analytical solution for an elastic half-space,  <i>(ii)</i> for elastic multilayered systems with the software package ELLEA [@Levenberg:2016a], and <i>(iii)</i> for viscoelastic multilayered systems with the software package ELLVA [@Levenberg:2016b]. Moreover, the code have been validated against benchmark results, covering a range of commonly used software packages, published by the European Commission [@amadeus:2000a]. The generic computational scheme proposed enables users to control any feature of the program, including geometrical and material properties, layer interface bonding, loading conditions and numerical parameters. Thus, ALVA supports advanced analysis, e.g., detailed analysis of surface tire interaction and unconventional axle loads, and enables users to balance numerical efficiency and accuracy. ALVA can be used as is [@Andersen:2020a;@Skar:2020a] or serve as computational kernel in new software; supporting future advances in pavement analysis software and development of model-guided data interpretation schemes.
 Future developments in this package may include mathematical formulations to describe more complex phenomena, e.g., <i>(i)</i> fragmented layer conditions (multi-cracked layers); <i>(ii)</i> both isotropic or transversely isotropic properties; <i>(iii)</i> both vertical and horizontal surface loads. 
 
 # References
