+++ README.md
@@ -101,7 +101,7 @@ The following examples are included in this package:
 
 | Software    |   R1    |     R2   |    R3   |    R4   |  R5   |     R6   |  R7     |    R8  |
 |-------------|---------|----------|---------|---------|-------|----------|---------|--------|
-| BISAR       |   0.7   |      0   | 186     |  170    |0      |     0    |  182    |  177   |
+| BISAR       |   0.7   |    N/A   | 186     |  170    |0      |   N/A    |  182    |  177   |
 | KENLAYER    |   1.5   |    -85   | 186     |  170    |0      |   -89    |  183    |  177   |
 | GAMES       |   0.7   |    -85   | 186     |  170    |0      |   -89    |  183    |  177   |
 | ALVA (slip) |   0.7   |  -84.9   | 186     | 169.7   | 0     |  -88.8   | 183.2   | 177.2  |
@@ -112,7 +112,7 @@ The following examples are included in this package:
 
 | Software    |   R1    |     R2   |    R3   |    R4   |  R5   |     R6   |  R7     |    R8  |
 |-------------|---------|----------|---------|---------|-------|----------|---------|--------|
-| BISAR       |   0.7   |       0  |     9   |   193   | 0     |     0    |  -12    |   204 |
+| BISAR       |   0.7   |     N/A  |     9   |   193   | 0     |   N/A    |  -12    |   204 |
 | KENLAYER    |   0.7   |    -120  |    -1   |   216   | 0     |   -78    |  -10    |   205 |
 | GAMES       |   0.7   |    -101  |    -3   |   194   | 0     |  -106    |    1    |   205 |
 | ALVA (slip) |   0.7   |  -102.6  |  -9.2   | 193.4   | 0     |  -107    |-11.5    | 204.1 |
