+++ paper.bib
@@ -112,7 +112,7 @@ structures},
   author = {Levenberg, Eyal and Skar, Asmus}
 }
 
-@article{maina:2004a,
+@article{Maina:2004a,
   title = {Developing software for elastic analysis of pavement structure responses to vertical and horizontal surface loadings},
   language = {eng},
   publisher={SAGE Publications Sage CA: Los Angeles, CA},
