### 2020-10-09 f82687e
Update ALVA_bonding_validation1.m
