### 2020-10-09 06f4076
Update ALVA_visco_validation1.m
