### 2020-10-09 93da21f
Update ALVA_let_validation3.m
