### 2020-10-09 d80fa54
Update ALVA_let_validation2.m
