### 2020-10-09 dc3d70d
Update ALVA_let_validation2.m
