### 2020-10-09 b402e2d
Update ALVA_let_validation1.m
