+++ README.md
@@ -4,7 +4,7 @@ Skar, Asmus, Andersen, Sebastian and Julius Nielsen (2020): Adaptive Layered Vis
 Adaptive Layered Viscoelastic Analysis (ALVA) is a MATLAB-package for pavement modeling. The aim with the software is to equip the civil engineering community with an advanced pavement modeling tool and computer package that is highly adaptive, transparent and open-access, capable of supporting current and future pavement evaluation needs. 
 
 <div>
-<img src="images/Roadmap_ALVA.png" width="90%">
+<img src="images/Roadmap_ALVA.png" width="100%">
 </div>
 
 The components of the ALVA software (shown in the Figure) is briefly described below:
@@ -44,7 +44,9 @@ ALVA comes with six validation examples (i.e., `main.m` scripts) comparing ALVA 
 The example files can be found in the in the _../ALVA/examples_ folder and the results obtained with independent codes can be found in the found in the _../ALVA/validation_ folder.  
 The following examples are included in this package:
 
-### Case 1 - Basic vailidation at critical points in the pavemenet (benchmark results)
+### Case 1 - Basic vailidation at critical points in the pavemenet system (benchmark results)
+In this section a basic validation of the ALVA package is presented, comparing ALVA repsonses to the [Advanced Models for Analytical Design of European Pavement Structures (AMADEUS, 2000)](https://trimis.ec.europa.eu/project/advanced-models-analytical-design-european-pavement-structures) report for benchmarking. As part of the AMADEUS study a number of popular response models were compared against the standard pavement structure shown in <b>Table 1</b>. Three packages, [BISAR](https://trid.trb.org/view/683379), [KENLAYER](https://www.webpages.uidaho.edu/ce475/files/software%20files/software.htm) and [GAMES](http://www.jsce.or.jp/committee/pavement/downloads/), that can facilitate layer interface slip situations, were further compared against ALVA for the proposed reference pavement system.
+
 * 'ALVA_bonding_validation1.m' - tests the implementation of the interface spring model, as well as compares the ALVA model with a range of commonly used softaware at critical positions within the pavament system published by the [European Commission](https://trimis.ec.europa.eu/project/advanced-models-analytical-design-european-pavement-structures).
 
 
@@ -53,7 +55,6 @@ The following examples are included in this package:
 | 1        |   260          |     5000            |   0.35           |
 | 2        |   500          |      200            |   0.40           |
 | 3        |   &infin;      |       50            |   0.45           |
-
 <p>
 <b>Table 1</b>: Reference pavement system used in basic validation.
 </p>
@@ -99,7 +100,6 @@ The following examples are included in this package:
 | KENLAYER    |   0.7   |    -120  |    1    |   216   |   0    |   -78   |  -10    |  205 |
 | GAMES       |   0.7   |    -119  |   11    |   217   |   0    |   -77   |   -1    |  205 |
 | ALVA (slip) |   0.7   |  -119.5  |  0.7    | 216.5   | 0.3    | -77.8   | -9.7    |204.7 |
-
 <p>
 <b>Table 4</b>: Reference pavement system subjeceted to a single wheel load - unbonded interface between layer 1 and 2, bonded interface between layer 2 and 3.
 </p>
@@ -117,11 +117,10 @@ The following examples are included in this package:
 
 | Software    |   R1    |     R2   |    R3   |    R4   |  R5   |     R6   |  R7     |    R8  |
 |-------------|---------|----------|---------|---------|-------|----------|---------|--------|
[…]
