+++ README.md
@@ -64,7 +64,7 @@ In this section a basic validation of the ALVA package is presented, comparing A
 | Dual (1) |   106.6        |     0.7           |   -170          |   0              |
 | Dual (2) |   106.6        |     0.7           |    170          |   0              |
 <p>
-<b>Table 2</b>: Load cases used in basic validation. 
+<b>Table 2</b>: Load cases used for basic validation of the code. 
 </p>
 
 | Name        | Response                                   | Location                   | Unit   |
@@ -79,7 +79,7 @@ In this section a basic validation of the ALVA package is presented, comparing A
 | R7          | Vertical strain at the top of layer 2      | ''                           | micron |
 | R8          | Vertical strain at the top of layer 3      | ''                           | micron |
 <p>
-<b>Table 2</b>: Description of response key-points used in basic validation.
+<b>Table 2</b>: Description of key-point responses used for basic validation of the code. 
 </p>
 
 | Software    |   R1    |     R2   |    R3   |    R4   |  R5   |     R6   |  R7     |    R8  |
@@ -87,8 +87,8 @@ In this section a basic validation of the ALVA package is presented, comparing A
 | BISAR       |   0.7   |  -100.5  |  251.7  |    185  |  0.4  |   -61.9  |  192.2  |  177.5 |
 | KENLAYER    |   0.8   |  -100.5  |  251.6  |  185.3  |  0.3  |     -62  |  192.2  |    177 |
 | GAMES       |   0.7   |  -100.5  |  251.6  |  185.1  |  0.3  |   -61.9  |  192.2  |  177.5 |
-| ALVA (slip) |   0.7   |  -100.4  |  251.6  |  185.1  |  0.3  |   -61.8  |  192.3  |  177.5 |
-| ALVA (bond) |   0.7   |  -100.4  |  251.6  |  185.1  |  0.3  |   -61.8  |  192.3  |  177.5 |
+| ALVA ('Slip') |   0.7   |  -100.4  |  251.6  |  185.1  |  0.3  |   -61.8  |  192.3  |  177.5 |
+| ALVA ('Bonded') |   0.7   |  -100.4  |  251.6  |  185.1  |  0.3  |   -61.8  |  192.3  |  177.5 |
 <p>
 <b>Table 3</b>: Reference pavement system subjeceted to a single wheel load - bonded interfaces.
 </p>
@@ -98,7 +98,7 @@ In this section a basic validation of the ALVA package is presented, comparing A
 | BISAR       |   0.7   |    -120  |    1    |   217   | 0.4    |   -78   |  -10    |  205 |
 | KENLAYER    |   0.7   |    -120  |    1    |   216   |   0    |   -78   |  -10    |  205 |
 | GAMES       |   0.7   |    -119  |   11    |   217   |   0    |   -77   |   -1    |  205 |
-| ALVA (slip) |   0.7   |  -119.5  |  0.7    | 216.5   | 0.3    | -77.8   | -9.7    |204.7 |
+| ALVA ('Slip') |   0.7   |  -119.5  |  0.7    | 216.5   | 0.3    | -77.8   | -9.7    |204.7 |
 <p>
 <b>Table 4</b>: Reference pavement system subjeceted to a single wheel load - unbonded interface between layer 1 and 2, bonded interface between layer 2 and 3.
 </p>
@@ -108,8 +108,8 @@ In this section a basic validation of the ALVA package is presented, comparing A
 | BISAR       |   0.7   |    N/A   | 186     |  170    |0      |   N/A    |  182    |  177   |
 | KENLAYER    |   1.5   |    -85   | 186     |  170    |0      |   -89    |  183    |  177   |
[…]
