+++ README.md
@@ -29,9 +29,70 @@ The core algorithm behind this package is based on LET, i.e., the classic formul
  <b>Figure 1: </b><i>N</i>-layered half-space model 
 </p>
 
-In this model all layers are assumed linear elastic, isotropic, homogeneous, fully bonded, and weightless. The model inputs include Young’s modulus E<sub>n</sub>, Poisson’s ratio <i>&Nu;<sub>n</sub></i>, and layer thickness t<sub>n</sub> (where <i>n</i> denotes the layer number). This model is engaged to calculate the response at any point, A<sub>j</sub>, of interest and for a given set of uniformly distributed circular loadings with load radius, <i>a</i>, and pressure <i>q</i>). An overview of LET model assumptions and solution procedure is given in [Khazanovich and Wang (2007)](https://journals.sagepub.com/doi/abs/10.3141/2037-06)
+In this model all layers are assumed linear elastic, isotropic, homogeneous, fully bonded, and weightless. The model inputs include Young’s modulus E<sub>n</sub>, Poisson’s ratio <i>&Nu;<sub>n</sub></i>, and layer thickness t<sub>n</sub> (where <i>n</i> denotes the layer number). This model is engaged to calculate the response at any point, A<sub>j</sub>, of interest and for a given set of uniformly distributed circular loadings with load radius, <i>a</i>, and pressure <i>q</i>). An overview of LET model assumptions and solution procedure is given in [Khazanovich and Wang (2007)](https://journals.sagepub.com/doi/abs/10.3141/2037-06).
 
-The viscoelastic response is approximated based on the LET calculations utilizing the methodology and load scheme suggested by [Levenberg (2016)](https://orbit.dtu.dk/en/publications/viscoelastic-pavement-modeling-with-a-spreadsheet) (see <b>Figure 2</b>). Viscoelastic layers are associated with a creep compliance of the form [Smith (1971)](https://onlinelibrary.wiley.com/doi/abs/10.1002/polc.5070350105?casa_token=TpvXuLD4yg8AAAAA:TcBBkISUJUuPIZN2fKHRLHHmNavv1OKzu1LKCWZ51C8_wCJto9Tn_ETbQht9EjOxkuEShFa-kxDZB5v8):
+* Input script
+``` % Select response analysis type
+alva.analysis = 'Full';
+% 1) 'Full'     : Conventional full integration with one-step Richardson
+%                 extrapolation (for improved convergence near surface) 
+%                 applied for evaluation of displacements and stresses
+% 2) 'PolFit'   : Use polynomial fit technique to reduce integral length
+%                 according to Andersen et al. (2018) for evaluation of
+%                 surface displacements
+``` 
+
+``` % Select interface 
+alva.bond = 'Bonded';
+% 1) 'Bonded'       : Full bonding between layers
+% 2) 'Slip'         : Interface bonding factor
+% 3) 'Frictionless' : No bonding between layers
+```
+% -------------------------------------------------------------------------
+% Numerical parameters
+% -------------------------------------------------------------------------
+alva.N  = 300;  % Number of Bessel zero points in numerical integration
+alva.n  = 30;   % Number of Gauss points points between zero points.
+
+% -------------------------------------------------------------------------
+% Pavement material properties (minimum two layers required)
+% -------------------------------------------------------------------------
+alva.zi = [150 750];         % Depth of first n-1 layers from the 
+                             % surface [mm]: last z = inf, and should not 
+                             % be added NB: zi(i) > zi(i-1) > z(i-2)...
[…]
