+++ README.md
@@ -27,7 +27,7 @@ Examples on ‘main.m’ scripts can be found in the _../ALVA/examples_ folder a
 The core algorithm behind this package is based on LET, i.e., the classic formulation for an <i>N</i>-layered half-space, shown in <b>Figure 2</b> 
 
 <div>
-<img src="images/N_layer.png" width="90%">
+<img src="images/N_layer.png" width="75%">
 </div>
 <p>
  <b>Figure 2:</b> <i>N</i>-layered half-space model 
@@ -35,7 +35,7 @@ The core algorithm behind this package is based on LET, i.e., the classic formul
 
 In this model all layers are assumed linear elastic, isotropic, homogeneous, fully bonded, and weightless. The model inputs include Young’s modulus <i>E<sub>n</sub></i>, Poisson’s ratio <i>&Nu;<sub>n</sub></i>, and layer thickness <i>t<sub>n</sub></i> (where <i>n</i> denotes the layer number). In addition ALVA allows users to define horizontally oriented springs, <i>k<sub>h</sub></i>, operating at the top or bottom layer interfaces to model imperfect interface conditions. This model is engaged to calculate the response at any point, <i>A<sub>j</sub></i>, of interest and for a given set of uniformly distributed circular loadings with load radius, <i>a</i>, and pressure <i>q</i>). An overview of LET model assumptions and solution procedure is given in [Khazanovich and Wang (2007)](https://journals.sagepub.com/doi/abs/10.3141/2037-06).
 
-* Response analysis type
+* Analysis type
 ``` 
 alva.analysis = 'Full';
 % 1) 'Full'     : Conventional full integration with one-step Richardson
@@ -48,7 +48,7 @@ alva.analysis = 'Full';
 * Interface
 
 <div>
-<img src="images/interface.png" width="90%">
+<img src="images/interface.png" width="75%">
 </div>
 <p>
  <b>Figure 3:</b> Close-up of interface showing a mechanical representation of the horizontal springs <i>k<sub>h</sub></i>
@@ -63,7 +63,7 @@ alva.bond = 'Bonded';
 * Numerical parameters
 
 <div>
-<img src="images/bessel1.png" width="90%">
+<img src="images/bessel1.png" width="75%">
 </div>
 <p>
  <b>Figure 4:</b> Visualisation of numerical parameters, i.e., Bessel function zero points 'N' and Gauss points 'n' for solving the inverse Hankel transform. In the Figure a Bessel function of the first kind of order zero is plotted and zero points are shown as red circles and gauss points as blue circles for the first interval and green circles for the second interval for 'N'='n'=5.
@@ -87,7 +87,7 @@ alva.kh = [1e6 1e6];         % Interface bonding/horizontal spring [MPa/mm]
 * Load configuration - Example dual load 
 
 <div>
-<img src="images/load.png" width="90%">
+<img src="images/load.png" width="75%">
 </div>
 <p>
  <b>Figure 5:</b> example dual wheel load.
@@ -137,20 +137,47 @@ epsxz = alva.epsxz.*1e6;
 ``` 
 
 ### Layered viscoelastic analysis
[…]
