+++ README.md
@@ -33,7 +33,7 @@ The core algorithm behind this package is based on LET, i.e., the classic formul
  <b>Figure 2:</b> <i>N</i>-layered half-space model 
 </p>
 
-In this model all layers are assumed linear elastic, isotropic, homogeneous, fully bonded, and weightless. The model inputs include Young’s modulus <i>E<sub>n</sub></i>, Poisson’s ratio <i>&Nu;<sub>n</sub></i>, and layer thickness <i>t<sub>n</sub></i> (where <i>n</i> denotes the layer number). In addition ALVA allows users to define horizontally oriented springs, <i>k<sub>h</sub></i>, operating at the top or bottom layer interfaces to model imperfect interface conditions. This model is engaged to calculate the response at any point, <i>A<sub>j</sub></i>, of interest and for a given set of uniformly distributed circular loadings with load radius, <i>a</i>, and pressure <i>q</i>). An overview of LET model assumptions and solution procedure is given in [Khazanovich and Wang (2007)](https://journals.sagepub.com/doi/abs/10.3141/2037-06).
+In this model all layers are assumed linear elastic, isotropic, homogeneous and weightless. The model inputs include Young’s modulus <i>E<sub>n</sub></i>, Poisson’s ratio <i>&Nu;<sub>n</sub></i>, and layer thickness <i>t<sub>n</sub></i> (where <i>n</i> denotes the layer number). In addition ALVA allows users to define horizontally oriented springs, <i>k<sub>h</sub></i>, operating at the top or bottom layer interfaces to model imperfect interface conditions. This model is engaged to calculate the response at any point, <i>A<sub>j</sub></i>, of interest and for a given set of uniformly distributed circular loadings with load radius, <i>a</i>, and pressure <i>q</i>). An overview of LET model assumptions and solution procedure is given in [Khazanovich and Wang (2007)](https://journals.sagepub.com/doi/abs/10.3141/2037-06).
 
 * Analysis type
 ``` 
@@ -56,9 +56,12 @@ alva.analysis = 'Full';
 
 ``` 
 alva.bond = 'Bonded';
-% 1) 'Bonded'       : Full bonding between layers
-% 2) 'Slip'         : Interface bonding factor
-% 3) 'Frictionless' : No bonding between layers
+% 'Bonded'       : Full bonding between layers
+% 'Slip'         : Interface bonding factor
+% 'Frictionless' : No bonding between layers
+
+ % Interface bonding/horizontal spring [MPa/mm]
+alva.kh = [1e6 1e6];        
 ```
 * Numerical parameters
 
@@ -73,7 +76,7 @@ alva.bond = 'Bonded';
 alva.N  = 300;  % Number of Bessel zero points in numerical integration
 alva.n  = 30;   % Number of Gauss points points between zero points.
 ``` 
-* Pavement material properties (minimum two layers required) - See  <b>Figure 1</b> 
+* Pavement material properties (minimum two layers required) - see <b>Figure 1</b> 
 
 ``` 
 alva.zi = [200 760];         % Depth of first n-1 layers from the 
@@ -81,16 +84,16 @@ alva.zi = [200 760];         % Depth of first n-1 layers from the
                              % be added NB: zi(i) > zi(i-1) > z(i-2)...
[…]
