### 2020-10-26 #2548 @AnaWaldila
@asmusskar, thank you to modify all topics of revision. Only 3 other topics I need that you modify, but in this case, are details.

- In repository <https://github.com/asmusskar/ALVA> and in paper <https://github.com/asmusskar/ALVA/blob/master/paper.md>, the topic "Main input parameters" presents the Figure 2 and their description.  This descritption presents "Poisson’s ratio Νn" and Figure 2 shows "ν". I understand "N" is the \nu greek letter, but I think in this case is better to use "ν" in description.
- In topic "Code Validation", there are two "Table 2" (Table 2: Load cases used for basic validation of the code.; Table 2: Description of key-point responses used for basic validation of the code.). Please, modify.
- Finally, topic "User examples" has Table 7. The  "Youngs moduli (MPa)" column shows "?". I saw your code ALVA_let_backcalculation.m and the initial parameter of Youngs Moduli is 200 in three layers. Is the Table 7 (column Youngs Moduli (MPa) correct in this case? Is not, in this case, to put this value (200) in column?

@Kevin-Mattheus-Moerman , after these modifications, my revision is finished. The author modifyed all recomendations and the paper, to me, is clear, well written, good examples and validations.
