+++ README.md
@@ -33,7 +33,7 @@ The core algorithm behind this package is based on LET, i.e., the classic formul
  <b>Figure 2:</b> <i>N</i>-layered half-space model 
 </p>
 
-In this model all layers are assumed linear elastic, isotropic, homogeneous and weightless. The model inputs include Young’s modulus <i>E<sub>n</sub></i>, Poisson’s ratio <i>&Nu;<sub>n</sub></i>, and layer thickness <i>t<sub>n</sub></i> (where <i>n</i> denotes the layer number). In addition ALVA allows users to define horizontally oriented springs, <i>k<sub>h</sub></i>, operating at the top or bottom layer interfaces to model imperfect interface conditions. This model is engaged to calculate the response at any point, <i>A<sub>j</sub></i>, of interest and for a given set of uniformly distributed circular loadings with load radius, <i>a</i>, and pressure <i>q</i>). An overview of LET model assumptions and solution procedure is given in [Khazanovich and Wang (2007)](https://journals.sagepub.com/doi/abs/10.3141/2037-06).
+In this model all layers are assumed linear elastic, isotropic, homogeneous and weightless. The model inputs include Young’s modulus <i>E<sub>n</sub></i>, Poisson’s ratio <i>&nu;<sub>n</sub></i>, and layer thickness <i>t<sub>n</sub></i> (where <i>n</i> denotes the layer number). In addition ALVA allows users to define horizontally oriented springs, <i>k<sub>h</sub></i>, operating at the top or bottom layer interfaces to model imperfect interface conditions. This model is engaged to calculate the response at any point, <i>A<sub>j</sub></i>, of interest and for a given set of uniformly distributed circular loadings with load radius, <i>a</i>, and pressure <i>q</i>). An overview of LET model assumptions and solution procedure is given in [Khazanovich and Wang (2007)](https://journals.sagepub.com/doi/abs/10.3141/2037-06).
 
 * Analysis type
 ``` 
@@ -258,7 +258,7 @@ In this section a basic validation of the ALVA package is presented, comparing A
 | R7          | Vertical strain at the top of layer 2      | ''                           | micron |
 | R8          | Vertical strain at the top of layer 3      | ''                           | micron |
 <p>
-<b>Table 2</b>: Description of key-point responses used for basic validation of the code. 
+<b>Table 3</b>: Description of key-point responses used for basic validation of the code. 
 </p>
 
 | Software    |   R1    |     R2   |    R3   |    R4   |  R5   |     R6   |  R7     |    R8  |
@@ -269,7 +269,7 @@ In this section a basic validation of the ALVA package is presented, comparing A
 | ALVA ('Slip') |   0.7   |  -100.4  |  251.6  |  185.1  |  0.3  |   -61.8  |  192.3  |  177.5 |
 | ALVA ('Bonded') |   0.7   |  -100.4  |  251.6  |  185.1  |  0.3  |   -61.8  |  192.3  |  177.5 |
 <p>
-<b>Table 3</b>: Reference pavement system subjeceted to a single wheel load - bonded interfaces.
+<b>Table 4</b>: Reference pavement system subjeceted to a single wheel load - bonded interfaces.
 </p>
 
 | Software    |   R1    |     R2   |    R3   |    R4   |  R5   |     R6   |  R7     |    R8  |
@@ -279,7 +279,7 @@ In this section a basic validation of the ALVA package is presented, comparing A
 | GAMES       |   0.7   |    -119  |   11    |   217   |   0    |   -77   |   -1    |  205 |
[…]
