### 2020-10-26 2aadb15
Update ALVA_visco_validation1.m
