### 2020-11-03 #2548 @AnaWaldila
@Kevin-Mattheus-Moerman I have a question, sorry for that. About the pdf paper. @asmusskar modified in repository the introduction and The presentation that has been in paper.md. Is The final pdf file made by the author or by the journal? Thanks!
