### 2020-11-05 #2548 @AnaWaldila
@Kevin-Mattheus-Moerman. I saw the final paper to check if it is allrigth. Sorry about that, but I did not understand why figures are not in pdf paper. I saw in this exact moment, the figures are in paper.md, like @asmusskar put.
