### 2020-11-07 #2548 @Kevin-Mattheus-Moerman
@openjournals/dev can you check why [this paper](https://github.com/openjournals/joss-papers/blob/joss.02548/joss.02548/10.21105.joss.02548.pdf) does not contain the images it shows in [the paper.md](https://github.com/asmusskar/ALVA/blob/master/paper.md). Is it because the images are in a different (`/images`) folder (so they are found on the repo but not by whedon?). They are using the following HTML format to render the images (which I assume is fine?):

```html
     <b>Figure 1</b>: N-Layered half-space model
```
