### 2020-11-09 #2548 @xuanxu
Figures should be included using markdown format with something similar to:
```
![N-layered half-space model.\label{fig:model}](images/N_layer.png)
```
([Example from the docs](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography))
