### 2020-11-09 6137f9e
Update paper.md
figures included
