+++ paper.md
@@ -28,7 +28,7 @@ To achieve this, some of the most promising numerical techniques that are curren
 
 ![$N$-layered half-space model.\label{fig:model}](images/N_layer.png){width=75%}
 
-In this model all layers are assumed linear elastic, isotropic, homogeneous, and weightless. The model inputs include Young’s modulus $E_{n}$, Poisson’s ratio $\nu_{n}$, and layer thickness $t_{n}$ (where $n$ denotes the layer number). This model is engaged to calculate the response (i.e stresses and defomations) at any point, $A_{j}$, of interest and for a given set of uniformly distributed circular loadings with load radius, $a$, and pressure $q$). In addition ALVA allows users to define horizontally oriented springs operating at the top or bottom layer interfaces to model imperfect interface conditions [@Levenberg:2020a]. Moreover, numerical features for improved code performance, e.g., acceleration techniques to speed computational time for analysis of surface displacements [@Andersen:2020a] and full response analysis [@khazanovich:2007a; @Levenberg:2016a], as well as extrapolation technique to improve convergence [@Erlingsson:2013a] was implemented. An overview of LET model assumptions and solution procedure is given in [@khazanovich:2007a].
+In this model all layers are assumed linear elastic, isotropic, homogeneous, and weightless. The model inputs include Young’s modulus $E_{n}$, Poisson’s ratio $\nu_{n}$, and layer thickness $t_{n}$ (where $n$ denotes the layer number). This model is engaged to calculate the response (i.e stresses and defomations) at any point, $A_{j}$, of interest and for a given set of uniformly distributed circular loadings with load radius, $a$, and pressure $q$). In addition ALVA allows users to define horizontally oriented springs operating at the top or bottom layer interfaces to model imperfect interface conditions [@Levenberg:2020a]. Moreover, numerical features for improved code performance are implemented, e.g., acceleration techniques to speed computational time for analysis of surface displacements [@Andersen:2020a] and full response analysis [@khazanovich:2007a; @Levenberg:2016a], as well as extrapolation technique to improve convergence [@Erlingsson:2013a]. An overview of LET model assumptions and solution procedure is given in [@khazanovich:2007a].
 
 The viscoelastic response is approximated based on the LET calculations utilizing the methodology and load scheme suggested by [@Levenberg:2016c] (see Figure \ref{fig:load}). Viscoelastic layers are associated with a creep compliance and model parameters $D_{0}$ and $D_{\infty}$, the short and long time compliances (respectively), and shape parameters $\tau_{D}$ and $n_{D}$, controlling the transition between $D_{0}$ and $D_{\infty}$. 
 
@@ -36,7 +36,8 @@ The viscoelastic response is approximated based on the LET calculations utilizin
 
 The load moves in a straight line from $x$=-$x_{0}$ (Start) to $x$=$x_{0}$ (End). The travel path is decomposed into $N$ intervals ($i$=1,…,$N$), each $\Delta x$ long. The point of response evaluation $A_{j}$ is indicated in the Figure; this point is located near the middle of the travel path (i.e., $x$-coordinate of zero), at $y$-coordinate $y_{0}$ and depth $z_{0}$ below the surface. 
 
[…]
