### 2020-11-10 #2548 @Kevin-Mattheus-Moerman
@asmusskar can you update your paper to use this markdown format for the images :point_up: ? You can run `@whedon generate pdf` when done.
