### 2020-11-16 #2548 @AnaWaldila
@asmusskar . Please, 2 cases to change:
"This model is engaged to calculate the response (i.e, stress and deformations). Please, change "deformations" to "strain".

"The point of response evaluation Aj is indicated in The Figure;". Aj is present in both Figures. Please, specify.

@Kevin-Mattheus-Moerman , thanks! After this two cases, I going to read The paper and I will accept. Do I need to check any box or complete any documentation? Do I only to informate to you my acception?
Thanks!
