### 2020-11-17 #2548 @CBenghi
@Kevin-Mattheus-Moerman,

I thought I could tick the check-list above but it appears disabled, even if I'm logged in; is there something I need to to before that?

![image](https://user-images.githubusercontent.com/5476273/99380559-6c468d80-28ca-11eb-96a9-fee76589bfeb.png)

Thanks.
