### 2020-11-17 3fc6cbd
Minor typos, richer first launch instructions.
