+++ README.md
@@ -36,7 +36,7 @@ The core algorithm behind this package is based on linear elastic theory (LET), 
 In this model all layers are assumed linear elastic, isotropic, homogeneous and weightless. The model inputs include Young’s modulus <i>E<sub>n</sub></i>, Poisson’s ratio <i>&nu;<sub>n</sub></i>, and layer thickness <i>t<sub>n</sub></i> (where <i>n</i> denotes the layer number). In addition ALVA allows users to define horizontally oriented springs, <i>k<sub>h</sub></i>, operating at the top or bottom layer interfaces to model imperfect interface conditions. This model is engaged to calculate the response at any point, <i>A<sub>j</sub></i>, of interest and for a given set of uniformly distributed circular loadings with load radius, <i>a</i>, and pressure <i>q</i>). An overview of LET model assumptions and solution procedure is given in [Khazanovich and Wang (2007)](https://journals.sagepub.com/doi/abs/10.3141/2037-06).
 
 * Analysis type
-``` 
+``` matlab
 alva.analysis = 'Full';
 % 1) 'Full'     : Conventional full integration with one-step Richardson
 %                 extrapolation (for improved convergence near surface) 
@@ -54,7 +54,7 @@ alva.analysis = 'Full';
  <b>Figure 3:</b> Close-up of interface showing a mechanical representation of the horizontal springs <i>k<sub>h</sub></i>
 </p>
 
-``` 
+``` matlab
 alva.bond = 'Bonded';
 % 'Bonded'       : Full bonding between layers
 % 'Slip'         : Interface bonding factor
@@ -72,7 +72,7 @@ alva.kh = [1e6 1e6];
  <b>Figure 4:</b> Visualisation of numerical parameters, i.e., Bessel function zero points 'N' and Gauss points 'n' for solving the inverse Hankel transform. In the Figure a Bessel function of the first kind of order zero is plotted with 'N'='n'=5. Zero points are shown as red circles, and Gauss points as blue and green circles for the first and second interval, respectively. 
 </p>
 
-``` 
+``` matlab
 alva.N  = 300;  % Number of Bessel zero points in numerical integration
 alva.n  = 30;   % Number of Gauss points points between zero points.
 ``` 
@@ -96,7 +96,7 @@ alva.nu = [0.35 0.40 0.45];  % Layer Poisson's ratio [-]
  <b>Figure 5:</b> Load configuration example: dual wheel load.
 </p>
 
-``` 
+``` matlab
 alva.q  = [0.7
            0.7];         % Load pressure [MPa] (uniform vertical pressure)
 alva.a  = [106.6
@@ -108,7 +108,7 @@ alva.Xl = [-170 0
 
 Response at the centre of one tire load, i.e., <i>x</i>=170 and <i>y</i>=0 with varying depth <i>z</i>=(0,259.99,260.1,760.1), and midpoint between tires, i.e., <i>x</i>=0 and <i>y</i>=0 with varying depth <i>z</i>=(0,259.99,260.1,760.1)
 
-``` 
+``` matlab
 alva.Xd = [170 0 0; 170 0 259.99; 170 0 260.1; 170 0 760.1;
              0 0 0;   0 0 259.99;   0 0 260.1;   0 0 760.1];   
 ```  
@@ -129,7 +129,7 @@ Strains:
 
[…]
