### 2020-11-18 #2548 @CBenghi
@asmusskar,

The paper ticks the large majority of the requirements except a few minor glitches.

1) **Automated tests**: I've posted an issue on the repository to notify about an error thrown in Octave by the label API that prevents me from running the validation script(s); and you have a claim it should work on Octave (I'm running 5.2.0, which is the latest).
2) **Community guidelines**: I can't find clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
3) **Quality of writing**: the language is clear, but I think the paper would improve substantially if you identified sub-sections in its content (and perhaps rearranged the order of some statements). I don't have a strong opinion on this, I would accept as is if you insisted; this is just a suggestion for easier readability.

Finally, I'm not sure of a method to evaluate your performance claim around "near real time" as the term itself is vague and context dependent. Again, not a deal breaker, though.

Thanks,
Claudio
