### 2020-11-18 a2b3a34
Update ALVA_let_validation1.m
bug fix fontsize legend to make comp. with Octave
