### 2020-11-18 1135f6d
Update ALVA_let_validation2.m
bug fix fontsize legend to make comp. with Octave
