### 2020-11-18 02925d8
Update ALVA_let_validation3.m
bug fix fontsize legend to make comp. with Octave
