### 2020-11-18 ad76bbc
Update ALVA_let_validation4.m
bug fix fontsize to make comp. w/Octave
