### 2020-11-18 57313a0
Update ALVA_let_validation5.m
bug fix fontsize to make comp. w/Octave
