### 2020-11-18 c9fc29e
Update ALVA_visco_validation1.m
bug fix fontsize
