### 2020-11-18 #2548 @Kevin-Mattheus-Moerman
Thanks for raising those points @CBenghi

@asmusskar can you work on those :point_up: ?
