+++ paper.md
@@ -36,7 +36,7 @@ The viscoelastic response is approximated based on the LET calculations utilizin
 
 The load moves in a straight line from $x$=-$x_{0}$ (Start) to $x$=$x_{0}$ (End). The travel path is decomposed into $N$ intervals ($i$=1,…,$N$), each of length $\Delta x$. The point of response evaluation $A_{j}$ is indicated in Figure \ref{fig:load}; this point is located near the middle of the travel path (i.e., $x$-coordinate of zero), at $y$-coordinate $y_{0}$ and depth $z_{0}$ below the surface. 
 
-The computer package ALVA offers a near real-time solution for the history of response inside the system at any point of interest resulting from a moving load. The code has been validated by comparing ALVA responses at several positions within pavement system to analytical solutions and other commonly used software packages, i.e.: $(i)$ for a one-layer model with the analytical solution for an elastic half-space, $(ii)$ for elastic multilayered systems with the software package ELLEA [@Levenberg:2016a], and $(iii)$ for viscoelastic multilayered systems with the software package ELLVA [@Levenberg:2016b]. Moreover, the code has been validated against benchmark results, covering a range of commonly used software packages, published by the European Commission [@amadeus:2000a]. The generic computational scheme proposed enables users to control any feature of the program, including geometrical and material properties, layer interface bonding, loading conditions and numerical parameters. Thus, ALVA supports advanced analysis, e.g., detailed analysis of surface tire interaction and unconventional axle loads, and enables users to balance numerical efficiency and accuracy. ALVA can be used as is [@Andersen:2020a;@Skar:2020a] or serve as computational kernel in new software; supporting future advances in pavement analysis software and development of model-guided data interpretation schemes.
+The computer package ALVA offers a solution for the history of response inside the system at any point of interest resulting from a moving load. The code has been validated by comparing ALVA responses at several positions within pavement system to analytical solutions and other commonly used software packages, i.e.: $(i)$ for a one-layer model with the analytical solution for an elastic half-space, $(ii)$ for elastic multilayered systems with the software package ELLEA [@Levenberg:2016a], and $(iii)$ for viscoelastic multilayered systems with the software package ELLVA [@Levenberg:2016b]. Moreover, the code has been validated against benchmark results, covering a range of commonly used software packages, published by the European Commission [@amadeus:2000a]. The generic computational scheme proposed enables users to control any feature of the program, including geometrical and material properties, layer interface bonding, loading conditions and numerical parameters. Thus, ALVA supports advanced analysis, e.g., detailed analysis of surface tire interaction and unconventional axle loads, and enables users to balance numerical efficiency and accuracy. ALVA can be used as is [@Andersen:2020a;@Skar:2020a] or serve as computational kernel in new software; supporting future advances in pavement analysis software and development of model-guided data interpretation schemes.
 
 Future developments in this package may include mathematical formulations to describe more complex phenomena, e.g., $(i)$ fragmented layer conditions (multi-cracked layers); $(ii)$ both isotropic or transversely isotropic properties; $(iii)$ both vertical and horizontal surface loads. 
 
