### 2020-11-18 9ed95c1
Update ALVA_bonding_validation1.m
