+++ examples/ALVA_bonding_validation1.m
@@ -188,70 +188,67 @@ slip_dual = [0.700    0   9  193 0.0000      0  -12  204;
 
 
 % -------------------------------------------------------------------------
-% Table
+% Display results as matlab table (not compatible with Octave)
 % -------------------------------------------------------------------------
 
-Location = {'A1';'A2';'A3';'A4';'A5';'A6';'A7';'A8'};
-Description     = [{'Vertical stress surface at center of load'};...
-            {'Horizontal strain bottom layer 1 at center of load'};...
-            {'Vertical strain top layer 2 at center of load'};...
-            {'Vertical strain top layer 3 at center of load'};...
-            {'Vertical stress surface at edge of load'};...
-            {'Horizontal strain bottom layer 1 at edge of load'};...
-            {'Vertical strain top layer 2 at edge of load'};...
-            {'Vertical strain top layer 3 at edge of load'}];
-TableGuide = table(Description,'RowNames',Location)
+%Location = {'A1';'A2';'A3';'A4';'A5';'A6';'A7';'A8'};
+%Description     = [{'Vertical stress surface at center of load'};...
+%            {'Horizontal strain bottom layer 1 at center of load'};...
+%            {'Vertical strain top layer 2 at center of load'};...
+%            {'Vertical strain top layer 3 at center of load'};...
+%            {'Vertical stress surface at edge of load'};...
+%            {'Horizontal strain bottom layer 1 at edge of load'};...
+%            {'Vertical strain top layer 2 at edge of load'};...
+%            {'Vertical strain top layer 3 at edge of load'}];
+%TableGuide = table(Description,'RowNames',Location)
 
-
-CodeName1 = {'BISAR';'KENLAYER';'GAMES';'ALVA (slip)';'ALVA (bond)'};
-CodeName2 = {'BISAR';'KENLAYER';'GAMES';'ALVA (slip)'};
-
+%CodeName1 = {'BISAR';'KENLAYER';'GAMES';'ALVA (slip)';'ALVA (bond)'};
+%CodeName2 = {'BISAR';'KENLAYER';'GAMES';'ALVA (slip)'};
 
 % Single wheel bonded interfaces
-A1sb = [round(bond_single(:,1),1);round(sigz2s(1),1);round(sigz1s(1),1)];
-A2sb = [round(bond_single(:,2),1);round(epsx2s(2),1);round(epsx1s(2),1)];
-A3sb = [round(bond_single(:,3),1);round(epsz2s(3),1);round(epsz1s(3),1)];
-A4sb = [round(bond_single(:,4),1);round(epsz2s(4),1);round(epsz1s(4),1)];
-A5sb = [round(bond_single(:,5),1);round(sigz2s(5),1);round(sigz1s(5),1)];
-A6sb = [round(bond_single(:,6),1);round(epsx2s(6),1);round(epsx1s(6),1)];
-A7sb = [round(bond_single(:,7),1);round(epsz2s(7),1);round(epsz1s(7),1)];
[…]
