### 2020-11-18 dea59e5
Update ALVA_bonding_validation1.m
