+++ examples/ALVA_bonding_validation1.m
@@ -188,67 +188,80 @@ slip_dual = [0.700    0   9  193 0.0000      0  -12  204;
 
 
 % -------------------------------------------------------------------------
-% Display results as matlab table (not compatible with Octave)
+% Plot results (table function not compatible with Octave)
 % -------------------------------------------------------------------------
 
-%Location = {'A1';'A2';'A3';'A4';'A5';'A6';'A7';'A8'};
-%Description     = [{'Vertical stress surface at center of load'};...
-%            {'Horizontal strain bottom layer 1 at center of load'};...
-%            {'Vertical strain top layer 2 at center of load'};...
-%            {'Vertical strain top layer 3 at center of load'};...
-%            {'Vertical stress surface at edge of load'};...
-%            {'Horizontal strain bottom layer 1 at edge of load'};...
-%            {'Vertical strain top layer 2 at edge of load'};...
-%            {'Vertical strain top layer 3 at edge of load'}];
+Location = {'R1';'R2';'R3';'R4';'R5';'R6';'R7';'R8'};
+Description     = [{'Vertical stress surface at center of load'};...
+            {'Horizontal strain bottom layer 1 at center of load'};...
+            {'Vertical strain top layer 2 at center of load'};...
+            {'Vertical strain top layer 3 at center of load'};...
+            {'Vertical stress surface at edge of load'};...
+            {'Horizontal strain bottom layer 1 at edge of load'};...
+            {'Vertical strain top layer 2 at edge of load'};...
+            {'Vertical strain top layer 3 at edge of load'}];
 %TableGuide = table(Description,'RowNames',Location)
 
-%CodeName1 = {'BISAR';'KENLAYER';'GAMES';'ALVA (slip)';'ALVA (bond)'};
-%CodeName2 = {'BISAR';'KENLAYER';'GAMES';'ALVA (slip)'};
+TableGuide = [{'R1: Vertical stress surface at center of load'};...
+            {'R2: Horizontal strain bottom layer 1 at center of load'};...
+            {'R3: Vertical strain top layer 2 at center of load'};...
+            {'R4: Vertical strain top layer 3 at center of load'};...
+            {'R5: Vertical stress surface at edge of load'};...
+            {'R6: Horizontal strain bottom layer 1 at edge of load'};...
+            {'R7: Vertical strain top layer 2 at edge of load'};...
+            {'R8: Vertical strain top layer 3 at edge of load'}]
+
+CodeName1 = {'BISAR';'KENLAYER';'GAMES';'ALVA (slip)';'ALVA (bond)'}
+CodeName2 = {'BISAR';'KENLAYER';'GAMES';'ALVA (slip)'};
 
 % Single wheel bonded interfaces
[…]
