### 2020-11-18 #2548 @asmusskar
@Kevin-Mattheus-Moerman, yes

@CBenghi,
1. I have resolved the latest issue. I hope you can accept the "quick-fix" until I come up with a more beatified solution.
2. I have included a short "How to contribute" section in the Readme.. I'm relatively new in this forum, so any additional inputs are very welcome
3. I'm not sure what is meant by " sub-sections in its content".. you want me to add sub-sections in the paper, e.g. ##Introduction, ##Input parameters ##Summary?... The text "near-real time" was deleted as suggested
