### 2020-11-19 #2548 @CBenghi
Hello @asmusskar,
Ok on 1 and 2.
I've also added a PR to the paper with minor grammar changes.

On number 3: yes that's what I meant. I think readability would benefit from that, but I would accept as is if you think differently. Just let me know.
