### 2020-11-20 #2548 @asmusskar
@CBenghi, thanks for those changes. I have updated the paper, dividing into sections as proposed 👇
