### 2020-11-24 #2548 @Kevin-Mattheus-Moerman
> @Kevin-Mattheus-Moerman, please note that a typo in the title of the paper was changed so that the name of this thread is different from the paper/software title (i.e. the word "elastic" was omitted).

Thanks @asmusskar I updated the title of this issue accordingly
