### 2020-11-24 #2548 @Kevin-Mattheus-Moerman
@asmusskar here are the last few steps

- [ ] Comment, should `...serve as computational kernel ...` be changed to `...serve as a computational kernel...`?

- [ ] What is the most recent version tag to use for this work? Is it still `v1.0.0` or has it moved on?

- [ ] Can you archive a copy of the reviewed software on [Zenodo](https://zenodo.org/) (some find [these steps](https://guides.github.com/activities/citable-code/) useful)? (let us know if you change the version in this process). Please make sure that the archive metadata like the title, the author names, and author order, matches that of the JOSS paper. You may have to manually edit this.
