### 2020-11-24 #2548 @asmusskar
@Kevin-Mattheus-Moerman, addressing your points above  ☝️
- yes (and the paper was changed accordingly)
- I created a new post review release v1.2 for Zenodo
- the DOI is 10.5281/zenodo.4289275
