### 2020-11-25 #2548 @arfon
@AnaWaldila, @CBenghi - many thanks for your reviews here and to @Kevin-Mattheus-Moerman for editing this submission. JOSS simply wouldn't work without the volunteer efforts of awesome folks like yourselves! ✨
